Add user management support.

+++ twinkle-security/twinkle-authorization/src/main/java/com/twinkle/cloud/security/authorization/component/CustomUserDetailsService.java
@@ -38,6 +38,8 @@ public class CustomUserDetailsService implements UserDetailsService {
         User user = userService.getByUniqueId(uniqueId);
         log.info("load user by username :{}", user.toString());
 
+
+
         return new org.springframework.security.core.userdetails.User(
                 user.getUsername(),
                 user.getPassword(),
@@ -55,7 +57,7 @@ public class CustomUserDetailsService implements UserDetailsService {
      * @return
      */
     protected Set<GrantedAuthority> obtainGrantedAuthorities(User user) {
-        Set<Role> roles = roleService.queryUserRolesByUserId(user.getId());
+        Set<Role> roles = roleService.queryUserRolesByUserId((Long)user.getId());
         log.info("user:{},roles:{}", user.getUsername(), roles);
         return roles.stream().map(role -> new SimpleGrantedAuthority(role.getCode())).collect(Collectors.toSet());
     }
